# [Optimized]: Removed several bounds checks