--- a/README.md
+++ b/README.md
@@ -38,10 +38,10 @@ Benchmarks
 
 ```
 name              time/op
-Simple/ASCII-4     449ns ± 1%
-Simple/French-4    872ns ± 1%
-Simple/Nordic-4   1.74µs ± 1%
-Simple/Tibetan-4  1.34µs ± 1%
+Simple/ASCII-4     365ns ± 1%
+Simple/French-4    680ns ± 2%
+Simple/Nordic-4   1.33µs ± 2%
+Simple/Tibetan-4  1.15µs ± 2%
 
 name              alloc/op
 Simple/ASCII-4     96.0B ± 0%