--- a/go.mod
+++ b/go.mod
@@ -1,1 +1,3 @@
 module github.com/agnivade/levenshtein
+
+go 1.13