--- a/levenshtein.go
+++ b/levenshtein.go
@@ -43,10 +43,13 @@ func ComputeDistance(a, b string) int {
 
 	// init the row
 	x := make([]int, lenS1+1)
-	for i := 0; i <= lenS1; i++ {
+	for i := 0; i < len(x); i++ {
 		x[i] = i
 	}
 
+	// make a dummy bounds check to prevent the 2 bounds check down below.
+	// The one inside the loop is particularly costly.
+	_ = x[lenS1]
 	// fill in the rest
 	for i := 1; i <= lenS2; i++ {
 		prev := i